feat: refactor StartScreen to use a dropdown for getting started instructions
